Makes the term not dependant of the OS

+++ src/utils.rs
@@ -81,18 +81,18 @@ impl ExactSizeIterator for EscapeRDF {
     }
 }
 
-pub fn to_bytes(int: usize) -> [u8; size_of::<usize>()] {
+pub fn to_bytes(int: u64) -> [u8; size_of::<u64>()] {
     //TODO: remove when next rust version stabilize this method
     unsafe { transmute(int) }
 }
 
-pub fn from_bytes(bytes: [u8; size_of::<usize>()]) -> usize {
+pub fn from_bytes(bytes: [u8; size_of::<u64>()]) -> u64 {
     //TODO: remove when next rust version stabilize this method
     unsafe { transmute(bytes) }
 }
 
-pub fn from_bytes_slice(bytes: &[u8]) -> usize {
-    let mut buf = [0 as u8; size_of::<usize>()];
+pub fn from_bytes_slice(bytes: &[u8]) -> u64 {
+    let mut buf = [0 as u8; size_of::<u64>()];
     buf.copy_from_slice(bytes);
     from_bytes(buf)
 }
